tests/wpt/tests/html/canvas/offscreen/line-styles/2d.line.width.scaledefault.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.line.width.scaledefault</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.line.width.scaledefault</h1>
<p class="desc">Default lineWidth strokes are affected by scale transformations</p>


<script>
var t = async_test("Default lineWidth strokes are affected by scale transformations");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 100, 50);

  ctx.scale(50, 50);
  ctx.strokeStyle = '#0f0';
  ctx.moveTo(0, 0.5);
  ctx.lineTo(2, 0.5);
  ctx.stroke();

  _assertPixel(canvas, 25,25, 0,255,0,255);
  _assertPixel(canvas, 50,25, 0,255,0,255);
  _assertPixel(canvas, 75,25, 0,255,0,255);
  _assertPixel(canvas, 50,5, 0,255,0,255);
  _assertPixel(canvas, 50,45, 0,255,0,255);
  t.done();

});
</script>
